enh(Autocomplete): Use infinite scroll Autocomplete as default one for ConnectedAutocomplete (#253)
