+++ src/InputField/Select/Autocomplete/Connected/Multi.tsx
@@ -3,6 +3,7 @@ import MultiAutocompleteField from '../Multi';
 
 const MultiConnectedAutocompleteField = ConnectedAutocompleteField(
   MultiAutocompleteField,
+  true,
 );
 
 export default MultiConnectedAutocompleteField;
