+++ src/InputField/Select/Autocomplete/Connected/Single.tsx
@@ -3,6 +3,7 @@ import AutocompleteField from '..';
 
 const SingleConnectedAutocompleteField = ConnectedAutocompleteField(
   AutocompleteField,
+  false,
 );
 
 export default SingleConnectedAutocompleteField;
