+++ src/InputField/Select/Autocomplete/Connected/index.stories.tsx
@@ -1,50 +1,68 @@
-/* eslint-disable react-hooks/rules-of-hooks */
-
-import React from 'react';
+import * as React from 'react';
 
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
 
-import MultiConnectedAutocompleteField from './Multi';
 import SingleConnectedAutocompleteField from './Single';
+import MultiConnectedAutocompleteField from './Multi';
 
 import { SelectEntry } from '../..';
+import { buildListingEndpoint } from '../../../..';
 
 export default { title: 'InputField/Autocomplete/Connected' };
 
-const mockedAxios = new MockAdapter(axios, { delayResponse: 500 });
+const buildEntities = (from) => {
+  return Array(10)
+    .fill(0)
+    .map((_, index) => ({
+      id: from + index,
+      name: `Entity ${from + index}`,
+    }));
+};
 
-const options = [
-  { id: 0, name: 'First Entity' },
-  { id: 1, name: 'Second Entity' },
-  { id: 2, name: 'Third Entity' },
-];
+const buildResult = (page) => ({
+  result: buildEntities((page - 1) * 10),
+  meta: {
+    pagination: {
+      limit: 10,
+      page,
+      total: 40,
+    },
+  },
+});
 
 const baseEndpoint = 'endpoint';
-const getSearchEndpoint = (): string => `${baseEndpoint}/search`;
+const getEndpoint = (params): string =>
+  buildListingEndpoint({ baseEndpoint, params, searchOptions: ['host.name'] });
 
-mockedAxios.onGet(baseEndpoint).reply(200, options);
+const mockedAxios = new MockAdapter(axios, { delayResponse: 500 });
 
-mockedAxios.onGet(`${baseEndpoint}/search`).reply(
-  200,
-  options.map((option) => ({ ...option, name: `${option.name} searched` })),
-);
+mockedAxios
+  .onGet(
+    /endpoint\?page=\d+(?:&search={"\$or":\[{"host\.name":{"\$rg":".*"}}]})?/,
+  )
+  .reply((config) => {
+    return [
+      200,
+      buildResult(parseInt(config.url?.split('page=')[1][0] || '0', 10)),
+    ];
+  });
 
-export const multiWithThreeOptionsRetrieved = (): JSX.Element => (
-  <MultiConnectedAutocompleteField
-    label="Multi Connected Autocomplete"
-    baseEndpoint={baseEndpoint}
-    getSearchEndpoint={getSearchEndpoint}
+export const single = (): JSX.Element => (
+  <SingleConnectedAutocompleteField
+    label="Single Connected Autocomplete"
+    initialPage={1}
+    getEndpoint={getEndpoint}
     getOptionsFromResult={(result): Array<SelectEntry> => result}
     placeholder="Type here..."
   />
 );
 
-export const singleWithThreeOptionsRetrieved = (): JSX.Element => (
-  <SingleConnectedAutocompleteField
-    label="Single Connected Autocomplete"
-    baseEndpoint={baseEndpoint}
-    getSearchEndpoint={getSearchEndpoint}
+export const multi = (): JSX.Element => (
+  <MultiConnectedAutocompleteField
+    label="Multi Infinite Autocomplete"
+    initialPage={1}
+    getEndpoint={getEndpoint}
     getOptionsFromResult={(result): Array<SelectEntry> => result}
     placeholder="Type here..."
   />
