--- src/InputField/Select/Autocomplete/InfiniteScroll/Multi.tsx
@@ -1,9 +1,0 @@
-import InfiniteAutocompleteField from '.';
-import MultiAutocompleteField from '../Multi';
-
-const MultiInfiniteAutocompleteField = InfiniteAutocompleteField(
-  MultiAutocompleteField,
-  true,
-);
-
-export default MultiInfiniteAutocompleteField;
