--- src/InputField/Select/Autocomplete/InfiniteScroll/Single.tsx
@@ -1,9 +1,0 @@
-import InfiniteAutocompleteField from '.';
-import AutocompleteField from '..';
-
-const SingleInfiniteAutocompleteField = InfiniteAutocompleteField(
-  AutocompleteField,
-  false,
-);
-
-export default SingleInfiniteAutocompleteField;
