+++ src/index.ts
@@ -119,4 +119,5 @@ export {
 export { default as buildListingDecoder } from './api/buildListingDecoder';
 
 export { default as copyToClipboard } from './utils/copy';
+export { default as useIntersectionObserver } from './utils/useIntersectionObserver';
 export { default as ContentWithCircularLoading } from './ContentWithCircularProgress';
